requirements.txt
aiofiles==0.3.2
httptools==0.0.11
psutil==5.4.6
sanic==0.7.0
ujson==1.35
uvloop==0.11.1
websockets==5.0
py-cpuinfo==5.0.0
